move proposals to separate files
